+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-04-09T15:22:30.842Z",
+    "2026-04-09T16:36:01.331Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {